+++ xaloon-core/xaloon-core-api/src/main/java/org/xaloon/core/api/search/SearchRequest.java
@@ -27,35 +27,35 @@ public abstract class SearchRequest implements Serializable {
 	 */
 	private static final long serialVersionUID = 1L;
 
-	private int firstRow = -1;
+	private long firstRow = -1;
 
-	private int maxRowCount = -1;
+	private long maxRowCount = -1;
 
 	/**
 	 * @return start position to search for items. -1,0 - search from start position
 	 */
-	public int getFirstRow() {
+	public long getFirstRow() {
 		return firstRow;
 	}
 
 	/**
 	 * @param firstRow
 	 */
-	public void setFirstRow(int firstRow) {
+	public void setFirstRow(long firstRow) {
 		this.firstRow = firstRow;
 	}
 
 	/**
 	 * @return maximum count of items to return. -1 - return all
 	 */
-	public int getMaxRowCount() {
+	public long getMaxRowCount() {
 		return maxRowCount;
 	}
 
 	/**
 	 * @param maxRowCount
 	 */
-	public void setMaxRowCount(int maxRowCount) {
+	public void setMaxRowCount(long maxRowCount) {
 		this.maxRowCount = maxRowCount;
 	}
 }
